Simplify write operations and permissions
BREAKING CHANGE!

Distinguishing between inserts and updates didn't really make sense
and was starting to cause problems so I'm combining those two operations
into a single `save` function.

As part of this I simplified permissions to just be read, write, admin.
It used to be (read, create, update, delete, admin).  That was
unnecessarily detailed and wasn't fitting the simplistic nature of
peerstack.