feat: Add boxed data
This allows a user to encrypt data so it can only be opened by the target user.
This has been added to the notification functionality so notifications
are encrypted if they are sent through the signaling server or through
web-push.

This is a breaking change to notifications and possibly other parts of
the library since the user object is expected to have an additonal
public key.  Therefore I'll be bumping the major version number.